Extract `operation` submodule from `supported_operations` module
The `supported_operations` module is divided into 2 parts:

- loading, updating, managing the supported operation set, matching on
  this set, etc. basically managing the collection
- new `supported_operations::operation` contains the functions and types
  related to reading and parsing single operation files from the
  filesystem
