+++ crates/extensions/c8y_mapper_ext/src/converter.rs
@@ -11,11 +11,11 @@ use crate::error::MessageConversionError;
 use crate::json;
 use crate::operations;
 use crate::operations::OperationHandler;
-use crate::supported_operations::get_child_ops;
-use crate::supported_operations::Operation;
+use crate::supported_operations::operation::get_child_ops;
+use crate::supported_operations::operation::Operation;
+use crate::supported_operations::operation::ResultFormat;
 use crate::supported_operations::Operations;
 use crate::supported_operations::OperationsError;
-use crate::supported_operations::ResultFormat;
 use crate::supported_operations::SupportedOperations;
 use anyhow::anyhow;
 use anyhow::Context;
@@ -1674,7 +1674,7 @@ pub(crate) mod tests {
     use crate::actor::IdUploadResult;
     use crate::config::BridgeConfig;
     use crate::config::C8yMapperConfig;
-    use crate::supported_operations::ResultFormat;
+    use crate::supported_operations::operation::ResultFormat;
     use crate::Capabilities;
     use anyhow::Result;
     use assert_json_diff::assert_json_include;
